Removed shared_ptr of stream_info_impl
When a pointer of stream_info_impl i(stream_info_impl *) s passed to the constructor of stream_outlet, a shared_ptr of stream_info_impl was created inside the stream_outlet. 
During the destruction of the stream_outlet, the shared_ptr is destructed which leads to unwanted delete of the stream_info_impl object.